backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.6.1,<3.0.0
python-dotenv==1.0.0
PyYAML==6.0.1
crewai-tools
crewai
click
langchain>=0.1.0
langchain-community
opentelemetry-instrumentation-starlette
daytona-sdk